backend/requirements.txt
fastapi==0.121.2
uvicorn[standard]==0.22.0
sqlmodel==0.0.27
jinja2==3.1.6
python-multipart==0.0.6
aiofiles==23.1.0
pytest==9.0.1
httpx==0.24.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==1.3.1
